+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog Fab Manager
 
+## next release
+- TODO export availabilities (taiga#57)
+- TODO bug: calendar (github#59)
+- TODO bug: delete event (github#61)
+- ONGOING spaces reservation
+- [TODO DEPLOY] `rake db:migrate`
+ - [TODO DEPLOY] `rake db:seed`
 ## v2.4.10 2017 January 9
 
 - Optimized notifications system